fix: Написаны тесты для режиссеров и популярных фильмов, исправлен 1 … (#29)
* fix: Написаны тесты для режиссеров и популярных фильмов, исправлен 1 sql запрос

* fix: Изменен порядок сортировки для популярных фильмов по жанрам при одинаковом количестве лайков

* fix: Исправлен порядок вывода популярных фильмов, при одинаковом количестве лайков
